Close MVP confirmation usability and decimal ledger gaps
M5 confirmation had a prompt-only client flow that submitted fixed financial facts, and fractional attendance deductions could make the existing student ledger unreadable. The confirmation path now captures actual facts explicitly, validates non-negative decimal lesson counts, and preserves fractional lesson balances through the finance read model. M4–M6 views use the three supported locales, and the QA handoff defines one bounded M1–M6 release regression.

Constraint: Monetary values remain integer base-currency amounts; only lesson-count facts may be fractional

Rejected: Adding a new payment or settlement capability | outside approved MVP scope

Confidence: high

Scope-risk: moderate

Directive: Do not replace the confirmation form with fixed defaults or scan fractional lesson balances into integer fields

Tested: gofmt; focused Go lesson/finance tests; frontend i18n and confirmation tests; frontend production build

Not-tested: Independent GLM/Kimi browser regression and human UAT remain release-gate work

+++ backend/internal/app/finance/service.go
@@ -177,10 +177,10 @@ func (s *Service) CreatePayment(ctx context.Context, user httpserver.AuthUser, w
 	if err != nil {
 		return Payment{}, false, repository.ErrDatabase
 	}
-	if err := s.repo.InsertRechargeLedger(ctx, tx, p, balance+amount, lessons+w.LessonsAdded, user.UserID, w.Note); err != nil {
+	if err := s.repo.InsertRechargeLedger(ctx, tx, p, balance+amount, lessons+float64(w.LessonsAdded), user.UserID, w.Note); err != nil {
 		return Payment{}, false, repository.ErrDatabase
 	}
-	if err := s.repo.UpdateEnrollmentBalances(ctx, tx, w.EnrollmentID, balance+amount, lessons+w.LessonsAdded); err != nil {
+	if err := s.repo.UpdateEnrollmentBalances(ctx, tx, w.EnrollmentID, balance+amount, lessons+float64(w.LessonsAdded)); err != nil {
 		return Payment{}, false, repository.ErrDatabase
 	}
 	if err := s.repo.LockBaseCurrency(ctx, tx); err != nil {
@@ -255,10 +255,10 @@ func (s *Service) VoidPayment(ctx context.Context, user httpserver.AuthUser, id 
 	if !ok {
 		return Payment{}, ErrInvalidState
 	}
-	if err := s.repo.InsertVoidLedger(ctx, tx, p, balance-p.AmountBase, lessons-p.LessonsAdded, user.UserID, reason); err != nil {
+	if err := s.repo.InsertVoidLedger(ctx, tx, p, balance-p.AmountBase, lessons-float64(p.LessonsAdded), user.UserID, reason); err != nil {
 		return Payment{}, repository.ErrDatabase
 	}
-	if err := s.repo.UpdateEnrollmentBalances(ctx, tx, p.EnrollmentID, balance-p.AmountBase, lessons-p.LessonsAdded); err != nil {
+	if err := s.repo.UpdateEnrollmentBalances(ctx, tx, p.EnrollmentID, balance-p.AmountBase, lessons-float64(p.LessonsAdded)); err != nil {
 		return Payment{}, repository.ErrDatabase
 	}
 	if err := s.repo.InsertAudit(ctx, tx, user.UserID, "PAYMENT_VOID", requestID); err != nil {
